requirements.txt
torch==2.2.0
transformers==4.33.0
opencv-python==4.7.0.72
Pillow==9.5.0
pyttsx3==2.90
